splitup resource_storage_management
